+++ generators/java/sdk/src/main/java/com/fern/java/client/ClientGeneratorContext.java
@@ -1,7 +1,7 @@
 package com.fern.java.client;
 
 import com.fern.generator.exec.model.config.GeneratorConfig;
-import com.fern.ir.model.ir.IntermediateRepresentation;
+import com.fern.irV42.model.ir.IntermediateRepresentation;
 import com.fern.java.AbstractGeneratorContext;
 
 public final class ClientGeneratorContext