+++ seed/fastapi/exhaustive/no-custom-config/resources/types/resources/union/types/animal.py
@@ -8,7 +8,7 @@ import typing
 import typing_extensions
 
 from ......core.datetime_utils import serialize_datetime
-from ......core.pydantic_utilities import pydantic_v1
+from ......core.pydantic_utilities import deep_union_pydantic_dicts, pydantic_v1
 from .cat import Cat as resources_types_resources_union_types_cat_Cat
 from .dog import Dog as resources_types_resources_union_types_dog_Dog
 
@@ -56,8 +56,12 @@ class Animal(pydantic_v1.BaseModel):
         return super().json(**kwargs_with_defaults)
 
     def dict(self, **kwargs: typing.Any) -> typing.Dict[str, typing.Any]:
-        kwargs_with_defaults: typing.Any = {"by_alias": True, "exclude_unset": True, **kwargs}
-        return super().dict(**kwargs_with_defaults)
+        kwargs_with_defaults_exclude_unset: typing.Any = {"by_alias": True, "exclude_unset": True, **kwargs}
+        kwargs_with_defaults_exclude_none: typing.Any = {"by_alias": True, "exclude_none": True, **kwargs}
+
+        return deep_union_pydantic_dicts(
+            super().dict(**kwargs_with_defaults_exclude_unset), super().dict(**kwargs_with_defaults_exclude_none)
+        )
 
     class Config:
         extra = pydantic_v1.Extra.forbid