+++ seed/pydantic/exhaustive/pydantic-v2/src/seed/exhaustive/resources/types/resources/union/cat.py
@@ -6,6 +6,7 @@ import typing
 import pydantic.v1 as pydantic
 
 from .....core.datetime_utils import serialize_datetime
+from .....core.pydantic_utilities import deep_union_pydantic_dicts
 
 
 class Cat(pydantic.BaseModel):
@@ -17,8 +18,12 @@ class Cat(pydantic.BaseModel):
         return super().json(**kwargs_with_defaults)
 
     def dict(self, **kwargs: typing.Any) -> typing.Dict[str, typing.Any]:
-        kwargs_with_defaults: typing.Any = {"by_alias": True, "exclude_unset": True, **kwargs}
-        return super().dict(**kwargs_with_defaults)
+        kwargs_with_defaults_exclude_unset: typing.Any = {"by_alias": True, "exclude_unset": True, **kwargs}
+        kwargs_with_defaults_exclude_none: typing.Any = {"by_alias": True, "exclude_none": True, **kwargs}
+
+        return deep_union_pydantic_dicts(
+            super().dict(**kwargs_with_defaults_exclude_unset), super().dict(**kwargs_with_defaults_exclude_none)
+        )
 
     class Config:
         allow_population_by_field_name = True