+++ generators/python/sdk/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.0.1] - 2024-05-13
+## [2.1.0] - 2024-05-14
 
 - Feature: Add support for cursor and offset pagination.
 
@@ -64,6 +64,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   pager = pager.next_page()
   print(pager.items)
   ```
+
+## [2.0.1] - 2024-05-14
+
+- Fix: the python generator previously used `exclude_unset` on pydantic models, however this would remove defaulted values. This change updates this to only exclude none fields that were not required.
+
 
 ## [2.0.0] - 2024-05-09